requirements.txt
PyPubSub==3.3.0
coverage==3.7.1
mock==1.0.1
nose==1.3.6
nose-parameterized==0.3.5
